+++ app/components/CardEditor.vue
@@ -1,12 +1,35 @@
 <template>
     <div>
-        <ul v-if="editor" class="editor-toolbar flex gap-3 flex-wrap">
+        <ul
+            v-if="emojiSelect && editor"
+            class="editor-toolbar flex gap-2 flex-wrap items-center mb-1"
+        >
+            <li>
+                <button
+                    type="button"
+                    @click="emojiSelect = false"
+                    class="block hover:text-secondary"
+                    v-tooltip="$t('back')"
+                >
+                    <ArrowLeft class="size-5" />
+                </button>
+            </li>
+            <li v-for="item in emojiList">
+                <button class="block" @click="setEmoji(item.code)">
+                    {{ item.icon }}
+                </button>
+            </li>
+        </ul>
+        <ul
+            v-else-if="editor"
+            class="editor-toolbar flex gap-2 flex-wrap py-0.5 mb-1"
+        >
             <li>
                 <button
                     type="button"
                     @click="editor.chain().focus().toggleBold().run()"
                     :disabled="!editor.can().chain().focus().toggleBold().run()"
-                    class="hover:text-secondary"
+                    class="block hover:text-secondary"
                     :class="{ 'text-secondary': editor.isActive('bold') }"
                     v-tooltip="$t('editorBold')"
                 >
@@ -20,7 +43,7 @@
                     :disabled="
                         !editor.can().chain().focus().toggleItalic().run()
                     "
-                    class="hover:text-secondary"
+                    class="block hover:text-secondary"
                     :class="{ 'text-secondary': editor.isActive('italic') }"
                     v-tooltip="$t('editorItalic')"
                 >
@@ -34,7 +57,7 @@
                     :disabled="
                         !editor.can().chain().focus().toggleStrike().run()
                     "
-                    class="hover:text-secondary"
+                    class="block hover:text-secondary"
                     :class="{ 'text-secondary': editor.isActive('strike') }"
                     v-tooltip="$t('editorStrike')"
                 >
@@ -45,7 +68,7 @@
                 <button
                     type="button"
                     @click="editor.chain().focus().toggleBulletList().run()"
-                    class="hover:text-secondary"
+                    class="block hover:text-secondary"
                     :class="{ 'text-secondary': editor.isActive('bulletList') }"
                     v-tooltip="$t('editorBulletList')"
                 >
@@ -56,7 +79,7 @@
                 <button
                     type="button"
                     @click="editor.chain().focus().toggleOrderedList().run()"
-                    class="hover:text-secondary"
+                    class="block hover:text-secondary"
                     :class="{
                         'text-secondary': editor.isActive('orderedList'),
                     }"
@@ -69,7 +92,7 @@
                 <button
                     type="button"
                     @click="editor.chain().focus().toggleTaskList().run()"
-                    class="hover:text-secondary"
+                    class="block hover:text-secondary"
                     :class="{
                         'text-secondary': editor.isActive('taskList'),
                     }"
@@ -82,7 +105,7 @@
                 <button
                     type="button"
                     @click="addImage"
-                    class="hover:text-secondary"
+                    class="block hover:text-secondary"
                     :class="{ 'text-secondary': editor.isActive('image') }"
                     v-tooltip="$t('editorImage')"
                 >
@@ -93,11 +116,21 @@
                 <button
                     type="button"
                     @click="editor.chain().focus().toggleCodeBlock().run()"
-                    class="hover:text-secondary"
+                    class="block hover:text-secondary"
                     :class="{ 'text-secondary': editor.isActive('codeBlock') }"
                     v-tooltip="$t('editorCodeblock')"
                 >
                     <Code class="size-5" />
+                </button>
+            </li>
+            <li>
+                <button
+                    type="button"
+                    @click="emojiSelect = true"
+                    class="block hover:text-secondary"
+                    v-tooltip="$t('editorEmojis')"
+                >
+                    <Smile class="size-5" />
                 </button>
             </li>
         </ul>
@@ -114,15 +147,30 @@ import {
     FileImage,
     Code,
     ListChecks,
+    Smile,
+    ArrowLeft,
 } from "lucide-vue-next";
 import { useEditor, EditorContent } from "@tiptap/vue-3";
 import { Placeholder } from "@tiptap/extensions";
 import { TaskItem, TaskList } from "@tiptap/extension-list";
 import Image from "@tiptap/extension-image";
+import Emoji, { emojis } from "@tiptap/extension-emoji";
 import FileHandler from "@tiptap/extension-file-handler";
 import StarterKit from "@tiptap/starter-kit";
 
 const model = defineModel();
+const emojiSelect = ref(false);
+
+const emojiList = [
+    { icon: "👍", code: "+1" },
+    { icon: "👎", code: "-1" },
+    { icon: "😄", code: "smile" },
+    { icon: "😢", code: "cry" },
+    { icon: "😓", code: "sweat" },
+    { icon: "🎉", code: "party" },
+    { icon: "🔥", code: "fire" },
+    { icon: "🚀", code: "rocket" },
+];
 
 const editor = useEditor({
     content: model.value,
@@ -135,6 +183,10 @@ const editor = useEditor({
         }),
         Image.configure({
             allowBase64: true,
+        }),
+        Emoji.configure({
+            emojis: emojis,
+            enableEmoticons: true,
         }),
         FileHandler.configure({
             allowedMimeTypes: [
@@ -226,5 +278,10 @@ const uploadImage = async (file: File) => {
         return { error: `Error uploading image: {$error}` };
         // You might want to show an error message to the user here
     }
+};
+
+const setEmoji = (emojiCode) => {
+    editor.value.chain().focus().setEmoji(emojiCode).run();
+    emojiSelect.value = false;
 };
 </script>